More documentation. Separating classes into separate files.
