+++ Assets/Scripts/NewGame/Entities/GameElements.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 2b9b7f76ab0c05c458de7ba4eb8ec042
+folderAsset: yes
+timeCreated: 1467549795
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
