+++ Assets/Scripts/NewGame/Entities/GameElements/DeadZone.cs
