+++ Assets/Scripts/NewGame/Entities/GameElements/DefaultPaddle.cs
