+++ Assets/Scripts/NewGame/Entities/GameObjectEliminator.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7497e7db753f5e746b54e6e27d1f30ee
+timeCreated: 1467549962
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
