+++ Assets/Scripts/NewGame/Entities/GameObjectPositioner.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: e5ac5ce8e9bd1eb4cb1442c9548542c0
-timeCreated: 1467545519
+guid: 8c0cd51220d88844bbbbeececa42375e
+timeCreated: 1467549815
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
