+++ Assets/Scripts/NewGame/Entities/Rigidbody2DSpeeder.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a86aefb597cd61944a8d2e0d95d93fc4
+timeCreated: 1467549906
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
