+++ Assets/Scripts/NewGame/Entities/SpriteRendererColorTinter.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5c6cf838f7e06e041ac45fe56b427fcb
+timeCreated: 1467550027
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
